requirements.txt
asgiref==3.5.2
Django==4.0
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
